Conflicts:
	proxi-client/src/app/create-client/create-client.component.html
	src/main/java/ma/jit/proxibanque/entities/Compte.java
	src/main/java/ma/jit/proxibanque/entities/CompteCourant.java
	src/main/java/ma/jit/proxibanque/entities/CompteEpargne.java
	src/main/java/ma/jit/proxibanque/metier/ClientMetierImpl.java

+++ src/main/java/ma/jit/proxibanque/dao/CompteRepository.java
@@ -17,6 +17,6 @@ import ma.jit.proxibanque.entities.Compte;
  *
  */
 @Repository
-public interface CompteRepository extends JpaRepository<Compte,Long > {
+public interface CompteRepository extends JpaRepository<Compte,Integer > {
 
 }
